fix: add pages for deployments
